+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 LR(0) parser Change Log
 =======================
 
+1.2.0
+-----
+
+*   Add: method `areChildrenMatch()` to interface `VovanVE\parser\common\TreeNodeInterface`
+    with its implementations in classes `VovanVE\parser\common\Token`
+    and `VovanVE\parser\tree\NonTerminal`.
+
 1.1.0
 -----